ref: Replace builder, add orderUtil date formatter
