+++ src/main/java/com/dellin/mondoc/utils/OrderUtil.java
@@ -1,5 +1,6 @@
 package com.dellin.mondoc.utils;
 
+import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 import java.time.ZoneId;
 
@@ -14,5 +15,12 @@ public final class OrderUtil {
 	public static Date getParsedDate(String date) {
 		LocalDate dateParse = LocalDate.parse(date);
 		return Date.from(dateParse.atStartOfDay(ZoneId.systemDefault()).toInstant());
+	}
+	
+	public static String getFormattedDate(LocalDate date, String format) {
+		SimpleDateFormat formatter = new SimpleDateFormat(format);
+		Date from = Date.from(date.atStartOfDay(ZoneId.systemDefault()).toInstant());
+		
+		return formatter.format(from);
 	}
 }
